+++ src/main.js
@@ -10,6 +10,7 @@ export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
 export { default as Table } from './components/Table';
 export { default as TabNav } from './components/TabNav';
+export { default as Textarea } from './components/Textarea';
 export { default as Token } from './components/Token';
 export { default as OverlayWrapper } from './components/OverlayWrapper';
 export { default as Popover } from './components/Popover';